Overhaul fallback logic
Fallback is now set up by configuring individual servers to build up a
fallback chain. Each server now only accepts a single fallback server
name and global fallback server definitions have been removed entirely.
Fallback is attempted if there is a kick or if the RUDP connection is
lost. If following the fallback chain is unsuccessful for any reason,
the client is disconnected with the original kick sent by the original
server or an error message about the loss of connection. If fallback
cannot be initiated at any part of the chain (e.g. due to the client not
being connected to a server in the first place) the client is
disconnected with an error message.

This commit fixes #150.