Log loaded plugins individually
The log messages say whether each successfully loaded plugin is
pre-built or automatically built by the proxy. In the former case the
file name is printed, in the latter case the directory name is printed
(not the path to the .so file contained within which is what is actually
loaded).

Closes #154.