Implement (inventory) formspec plugin API
Node-based formspecs are not supported because there is no node API yet.
Proxy-level formspec functionality is invisible to upstream servers. The
API can be used to display a formspec that is submitted to the upstream
server. It is also able to intercept the submission of formspecs
displayed by the upstream server.

Closes #131.