Fix the same username being able to have multiple sessions
Previously, a client failing this check would be disconnected,
incorrectly resulting in the username being removed from the playerlist
and enabling another client to log in. The only limit to this was the
number of internal servers accessible to the account. This commit fixes
the bypass by preventing the username from being removed from the list
in error cases.